refactor: move the details of fetching products out of the request handler
